backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
dnspython==2.6.1
beautifulsoup4==4.12.3
python-whois==0.9.4
reportlab==4.2.0
python-multipart==0.0.9
websockets==12.0
lxml==5.2.1
